backend/requirements.txt
fastapi>=0.115,<0.120
uvicorn[standard]>=0.32,<0.40
pydantic>=2.9,<3
pyyaml>=6.0
google-genai>=1.40,<2
mcp>=1.9,<2
httpx>=0.27,<0.30
sse-starlette>=2.1,<3
python-multipart>=0.0.12
